Add files via upload
Attached reg_file to a counter to slow down the clock cycles for testing purposes. Had to remove rst. Not sure if that is significant. 

TODO: test reg_file more thoroughly for (1) reading addresses is right (2) writing addresses is right (3) attachment to the rest of the infrastructure i.e. control and alu is correct.