Add files via upload
Attempt at a sample fsm control unit in ctl. Semi successful. 
Successful in implementing states and transition. 
However, a persistent bug is when B1 state constantly shows the LED for a1 to be lit even though not in A1 state (checked for rapid state change already and that is not the issue). As I am unable to isolate the problem, I shall move onto hardcoding the fsm for our actual implementation to see if the error gets replicated.